+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ------
 
+* New "actions" command to show ledger actions
 * Add ability to turn of fetching of trades, balances or actions for each account
 * "quote" subcommand now has a timestamp (-t) option to query historical prices